move addressTest to xrpl4j-model
