+++ xrpl4j-address-codec/src/test/java/org/xrpl/xrpl4j/codec/addresses/AddressCodecTest.java
@@ -177,7 +177,6 @@ public class AddressCodecTest {
     );
   }
 
-
   private void testEncodeDecode(
     Function<UnsignedByteArray, String> encoder,
     Function<String, UnsignedByteArray> decoder,
